Move brand designer into sub folder
